// src/generated/protocol/briosa/lifecycle.ts
export enum SpatialAnalyzerApplicationState {
  SPATIAL_ANALYZER_APPLICATION_STATE_UNSPECIFIED = 0,
  SPATIAL_ANALYZER_APPLICATION_STATE_NOT_RUNNING = 1,
  SPATIAL_ANALYZER_APPLICATION_STATE_STARTING = 2,
  SPATIAL_ANALYZER_APPLICATION_STATE_RUNNING = 3,
  SPATIAL_ANALYZER_APPLICATION_STATE_CLOSING = 4,
  SPATIAL_ANALYZER_APPLICATION_STATE_EXITED = 5,
  SPATIAL_ANALYZER_APPLICATION_STATE_AMBIGUOUS = 6,
  SPATIAL_ANALYZER_APPLICATION_STATE_FAULTED = 7,
  UNRECOGNIZED = -1,
}

export enum SpatialAnalyzerOwnership {
  SPATIAL_ANALYZER_OWNERSHIP_UNSPECIFIED = 0,
  SPATIAL_ANALYZER_OWNERSHIP_NONE = 1,
  SPATIAL_ANALYZER_OWNERSHIP_EXTERNAL = 2,
  SPATIAL_ANALYZER_OWNERSHIP_SERVER_LAUNCHED = 3,
  UNRECOGNIZED = -1,
}

export enum SpatialAnalyzerSdkState {
  SPATIAL_ANALYZER_SDK_STATE_UNSPECIFIED = 0,
  SPATIAL_ANALYZER_SDK_STATE_STOPPED = 1,
  SPATIAL_ANALYZER_SDK_STATE_STARTING = 2,
  SPATIAL_ANALYZER_SDK_STATE_RUNNING = 3,
  SPATIAL_ANALYZER_SDK_STATE_CONNECTING = 4,
  SPATIAL_ANALYZER_SDK_STATE_VERIFYING = 5,
  SPATIAL_ANALYZER_SDK_STATE_READY = 6,
  SPATIAL_ANALYZER_SDK_STATE_STOPPING = 7,
  SPATIAL_ANALYZER_SDK_STATE_RECOVERING = 8,
  SPATIAL_ANALYZER_SDK_STATE_FAULTED = 9,
  UNRECOGNIZED = -1,
}

export enum SpatialAnalyzerSdkRecoveryState {
  SPATIAL_ANALYZER_SDK_RECOVERY_STATE_UNSPECIFIED = 0,
  SPATIAL_ANALYZER_SDK_RECOVERY_STATE_NOT_REQUIRED = 1,
  SPATIAL_ANALYZER_SDK_RECOVERY_STATE_RECOVERY_AVAILABLE = 2,
  SPATIAL_ANALYZER_SDK_RECOVERY_STATE_OPERATOR_ACTION_REQUIRED = 3,
  SPATIAL_ANALYZER_SDK_RECOVERY_STATE_BLOCKED = 4,
  UNRECOGNIZED = -1,
}

export enum SpatialAnalyzerSdkTerminationKind {
  SPATIAL_ANALYZER_SDK_TERMINATION_KIND_UNSPECIFIED = 0,
  SPATIAL_ANALYZER_SDK_TERMINATION_KIND_START_FAILED = 1,
  SPATIAL_ANALYZER_SDK_TERMINATION_KIND_SDK_PROCESS_EXITED = 2,
  SPATIAL_ANALYZER_SDK_TERMINATION_KIND_SDK_CONNECTION_LOST = 3,
  SPATIAL_ANALYZER_SDK_TERMINATION_KIND_WORKER_PROCESS_EXITED = 4,
  SPATIAL_ANALYZER_SDK_TERMINATION_KIND_CONTROL_CHANNEL_LOST = 5,
  SPATIAL_ANALYZER_SDK_TERMINATION_KIND_WATCHDOG_TERMINATED = 6,
  UNRECOGNIZED = -1,
}
// ...
export interface SpatialAnalyzerLifecycleState {
  stateRevision?: bigint | undefined;
  applicationState?: SpatialAnalyzerApplicationState | undefined;
  ownership?: SpatialAnalyzerOwnership | undefined;
  applicationGeneration?: number | undefined;
  diagnosticCode?: string | undefined;
}
// ...
export interface SpatialAnalyzerSdkLifecycleState {
  stateRevision?: bigint | undefined;
  sdkState?: SpatialAnalyzerSdkState | undefined;
  sdkGeneration?: number | undefined;
  applicationGeneration?: number | undefined;
  connectionState?: SpatialAnalyzerConnectionState | undefined;
  executionReadinessState?: SpatialAnalyzerExecutionReadinessState | undefined;
  readyForMp?: boolean | undefined;
  recoveryState?: SpatialAnalyzerSdkRecoveryState | undefined;
  lastIncident?: SpatialAnalyzerSdkIncident | undefined;
  diagnosticCode?: string | undefined;
}

export interface SpatialAnalyzerSdkIncident {
  sdkGeneration?: number | undefined;
  terminationKind?: SpatialAnalyzerSdkTerminationKind | undefined;
  executionDisposition?: ExecutionDisposition | undefined;
  operationId?: string | undefined;
  diagnosticCode?: string | undefined;
}
// ...
export const SpatialAnalyzerLifecycleState: MessageFns<SpatialAnalyzerLifecycleState> = {
  encode(message: SpatialAnalyzerLifecycleState, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.stateRevision !== undefined && message.stateRevision !== 0n) {
      if (BigInt.asUintN(64, message.stateRevision) !== message.stateRevision) {
        throw new globalThis.Error("value provided for field message.stateRevision of type uint64 too large");
      }
      writer.uint32(8).uint64(message.stateRevision);
    }
    if (message.applicationState !== undefined && message.applicationState !== 0) {
      writer.uint32(16).int32(message.applicationState);
    }
    if (message.ownership !== undefined && message.ownership !== 0) {
      writer.uint32(24).int32(message.ownership);
    }
    if (message.applicationGeneration !== undefined) {
      writer.uint32(32).int32(message.applicationGeneration);
    }
    if (message.diagnosticCode !== undefined) {
      writer.uint32(42).string(message.diagnosticCode);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SpatialAnalyzerLifecycleState {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSpatialAnalyzerLifecycleState();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.stateRevision = reader.uint64() as bigint;
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.applicationState = reader.int32() as any;
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.ownership = reader.int32() as any;
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.applicationGeneration = reader.int32();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.diagnosticCode = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create(base?: DeepPartial<SpatialAnalyzerLifecycleState>): SpatialAnalyzerLifecycleState {
    return SpatialAnalyzerLifecycleState.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<SpatialAnalyzerLifecycleState>): SpatialAnalyzerLifecycleState {
    const message = createBaseSpatialAnalyzerLifecycleState();
    message.stateRevision = (object.stateRevision !== undefined && object.stateRevision !== null)
      ? BigInt(object.stateRevision)
      : 0n;
    message.applicationState = object.applicationState ?? 0;
    message.ownership = object.ownership ?? 0;
    message.applicationGeneration = object.applicationGeneration ?? undefined;
    message.diagnosticCode = object.diagnosticCode ?? undefined;
    return message;
  },
};
// ...
export const SpatialAnalyzerSdkIncident: MessageFns<SpatialAnalyzerSdkIncident> = {
  encode(message: SpatialAnalyzerSdkIncident, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.sdkGeneration !== undefined && message.sdkGeneration !== 0) {
      writer.uint32(8).int32(message.sdkGeneration);
    }
    if (message.terminationKind !== undefined && message.terminationKind !== 0) {
      writer.uint32(16).int32(message.terminationKind);
    }
    if (message.executionDisposition !== undefined) {
      writer.uint32(24).int32(message.executionDisposition);
    }
    if (message.operationId !== undefined) {
      writer.uint32(34).string(message.operationId);
    }
    if (message.diagnosticCode !== undefined) {
      writer.uint32(42).string(message.diagnosticCode);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SpatialAnalyzerSdkIncident {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSpatialAnalyzerSdkIncident();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.sdkGeneration = reader.int32();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.terminationKind = reader.int32() as any;
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.executionDisposition = reader.int32() as any;
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.operationId = reader.string();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.diagnosticCode = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create(base?: DeepPartial<SpatialAnalyzerSdkIncident>): SpatialAnalyzerSdkIncident {
    return SpatialAnalyzerSdkIncident.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<SpatialAnalyzerSdkIncident>): SpatialAnalyzerSdkIncident {
    const message = createBaseSpatialAnalyzerSdkIncident();
    message.sdkGeneration = object.sdkGeneration ?? 0;
    message.terminationKind = object.terminationKind ?? 0;
    message.executionDisposition = object.executionDisposition ?? undefined;
    message.operationId = object.operationId ?? undefined;
    message.diagnosticCode = object.diagnosticCode ?? undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | bigint | undefined;

type DeepPartial<T> = T extends bigint ? string | number | bigint
  : T extends Builtin ? T : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends { $case: string; value: unknown } ? { $case: T["$case"]; value?: DeepPartial<T["value"]> }
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}